pythonProject/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Session==0.5.0
rich==13.7.0
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
Werkzeug==2.3.7
psycopg2-binary==2.9.9
cachelib==0.10.2
